# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Farmart Grocery Store API",
    "description": "API Specification ออกแบบจากฟังก์ชันและหน้าเว็บทั้งหมดของระบบ Farmart Supermarket / E-Commerce (Next.js App Router)\n\nครอบคลุมทั้งระบบแค็ตตาล็อกสินค้า (Products), หมวดหมู่และแบรนด์ (Categories & Brands), ดีลพิเศษประจำวัน (Top Saver, Best Seller, Just Landing), ระบบรีวิวและการให้คะแนน (Reviews), ตะกร้าสินค้า (Cart), รายการสินค้าที่ชอบ (Wishlist), จดหมายข่าว (Newsletter), และสถานะระบบ (Health Check)",
    "version": "1.0.0",
    "contact": {
      "name": "Farmart Engineering Team",
      "email": "api-support@farmart.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "/",
      "description": "Current Origin Server"
    }
  ],
...
  "paths": {
...
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "ดึงข้อมูลรายละเอียดสินค้าเดี่ยว (Get Product Details)",
        "description": "ดึงข้อมูลรายละเอียดฉบับเต็มของสินค้า โดยระบุ `id` หรือ `slug` รวมถึงตารางโภชนาการ แหล่งที่มา และแกลเลอรีรูปภาพ",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID สินค้า หรือ Slug สินค้า",
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "พบข้อมูลสินค้า",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetail"
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบสินค้าตาม ID ที่ระบุ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}/related": {
      "get": {
        "tags": ["Products"],
        "summary": "ดึงรายการสินค้าที่เกี่ยวข้อง (Related Products)",
        "description": "ดึงสินค้าในหมวดหมู่เดียวกันหรือสินค้าแนะนำสำหรับหน้ารายละเอียดสินค้า",
        "operationId": "getRelatedProducts",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID สินค้าหลัก",
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "ดึงสินค้าที่เกี่ยวข้องสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductCardItem"
                  }
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบสินค้าหลัก",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/categories/{slug}": {
      "get": {
        "tags": ["Categories & Brands"],
        "summary": "ดึงข้อมูลหมวดหมู่ตาม Slug",
        "operationId": "getCategoryBySlug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "description": "Slug ของหมวดหมู่ เช่น fruits-vegetables",
            "schema": {
              "type": "string"
            },
            "example": "fruits-vegetables"
          }
        ],
        "responses": {
          "200": {
            "description": "พบหมวดหมู่",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบหมวดหมู่",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products/best-sellers": {
      "get": {
        "tags": ["Deals & Collections"],
        "summary": "สินค้าขายดี (Best Seller Collection)",
        "description": "ดึงรายการสินค้าขายดียอดนิยม รองรับการกรองตามแท็บหมวดหมู่ (All, Fruits & Vegetables, ฯลฯ)",
        "operationId": "getBestSellers",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "description": "กรองตามหมวดหมู่ (default: All)",
            "schema": {
              "type": "string",
              "default": "All"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ดึงสินค้าขายดีสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductCardItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/just-landing": {
      "get": {
        "tags": ["Deals & Collections"],
        "summary": "สินค้าเข้าใหม่ล่าสุด (Just Landing Collection)",
        "description": "ดึงรายการสินค้าใหม่ล่าสุดในร้านค้า",
        "operationId": "getJustLanding",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "All"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ดึงสินค้าใหม่สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductCardItem"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProductCardItem": {
        "type": "object",
        "required": ["id", "name", "price", "rating", "reviewCount"],
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Fresh Organic Hass Avocado (Pack of 4)"
          },
          "brand": {
            "type": "string",
            "example": "Farmart Organic Direct"
          },
          "category": {
            "type": "string",
            "example": "Fruits & Vegetables"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 6.49
          },
          "oldPrice": {
            "type": "number",
            "format": "float",
            "example": 8.99
          },
          "unit": {
            "type": "string",
            "example": "4 pcs (~650g)"
          },
          "discountPercent": {
            "type": "integer",
            "example": 28
          },
          "isOrganic": {
            "type": "boolean",
            "example": true
          },
          "rating": {
            "type": "number",
            "format": "float",
            "example": 4.9
          },
          "reviewCount": {
            "type": "integer",
            "example": 128
          },
          "imageUrl": {
            "type": "string",
            "example": "/images/products/avocado.png"
          }
        }
      },
      "NutritionItem": {
        "type": "object",
        "required": ["name", "amount"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Total Fat"
          },
          "amount": {
            "type": "string",
            "example": "12g"
          },
          "dailyValue": {
            "type": "string",
            "example": "15%"
          }
        }
      },
      "OriginInfo": {
        "type": "object",
        "required": ["farmName", "location", "harvestDate", "storageTemp", "shelfLife"],
        "properties": {
          "farmName": {
            "type": "string",
            "example": "Green Valley Regenerative Cooperative"
          },
          "location": {
            "type": "string",
            "example": "Chiang Mai Highlands, Thailand"
          },
          "harvestDate": {
            "type": "string",
            "example": "Yesterday morning (06:30 AM)"
          },
          "storageTemp": {
            "type": "string",
            "example": "12°C - 15°C (Cool Pantry)"
          },
          "shelfLife": {
            "type": "string",
            "example": "5 to 7 days from delivery date"
          }
        }
      },
      "GalleryItem": {
        "type": "object",
        "required": ["id", "label"],
        "properties": {
          "id": {
            "type": "string",
            "example": "front"
          },
          "label": {
            "type": "string",
            "example": "Fresh Whole Fruit"
          },
          "imageUrl": {
            "type": "string",
            "example": "/images/products/avocado-front.png"
          }
        }
      },
      "ReviewItem": {
        "type": "object",
        "required": ["id", "author", "rating", "date", "title", "comment", "verified", "helpfulCount"],
        "properties": {
          "id": {
            "type": "string",
            "example": "rev-1"
          },
          "author": {
            "type": "string",
            "example": "Sarah Jenkins"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "date": {
            "type": "string",
            "example": "2 days ago"
          },
          "title": {
            "type": "string",
            "example": "Best organic avocados I've ever ordered online!"
          },
          "comment": {
            "type": "string",
            "example": "Arrived in flawless condition with zero bruises. Cut one open immediately and the flesh was velvety and creamy."
          },
          "verified": {
            "type": "boolean",
            "example": true
          },
          "helpfulCount": {
            "type": "integer",
            "example": 24
          }
        }
      },
      "ProductDetail": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "brand",
          "category",
          "categorySlug",
          "price",
          "unit",
          "rating",
          "reviewCount",
          "stockStatus",
          "stockCount",
          "sku",
          "barcode",
          "shortDescription",
          "fullDescription",
          "highlights",
          "nutritionFacts",
          "originInfo",
          "gallery",
          "reviews"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "slug": {
            "type": "string",
            "example": "organic-hass-avocado"
          },
          "name": {
            "type": "string",
            "example": "Fresh Organic Hass Avocado (Pack of 4)"
          },
          "brand": {
            "type": "string",
            "example": "Farmart Organic Direct"
          },
          "category": {
            "type": "string",
            "example": "Fruits & Vegetables"
          },
          "categorySlug": {
            "type": "string",
            "example": "fruits-vegetables"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 6.49
          },
          "oldPrice": {
            "type": "number",
            "format": "float",
            "example": 8.99
          },
          "unit": {
            "type": "string",
            "example": "4 pcs (~650g)"
          },
          "discountPercent": {
            "type": "integer",
            "example": 28
          },
          "isOrganic": {
            "type": "boolean",
            "example": true
          },
          "rating": {
            "type": "number",
            "format": "float",
            "example": 4.9
          },
          "reviewCount": {
            "type": "integer",
            "example": 128
          },
          "stockStatus": {
            "type": "string",
            "enum": ["in_stock", "low_stock", "out_of_stock"],
            "example": "in_stock"
          },
          "stockCount": {
            "type": "integer",
            "example": 84
          },
          "sku": {
            "type": "string",
            "example": "FM-ORG-AVO-04"
          },
          "barcode": {
            "type": "string",
            "example": "8859123400192"
          },
          "shortDescription": {
            "type": "string",
            "example": "Hand-picked buttery Hass avocados cultivated on certified regenerative organic orchards."
          },
          "fullDescription": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Our Fresh Organic Hass Avocados are nurtured under temperate sun drenched orchards without any synthetic pesticides.",
              "Avocados are known as nature's superfood, loaded with heart-healthy monounsaturated oleic acid, fiber, and potassium."
            ]
          },
          "highlights": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "100% Certified Organic (USDA & Bio-Organic)",
              "Zero Synthetic Chemical Sprays or Waxes",
              "Picked within 36 hours of your delivery"
            ]
          },
          "nutritionFacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NutritionItem"
            }
          },
          "originInfo": {
            "$ref": "#/components/schemas/OriginInfo"
          },
          "gallery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GalleryItem"
            }
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewItem"
            }
          },
          "relatedIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["2", "3", "4"]
          }
        }
      },
...
      "Category": {
        "type": "object",
        "required": ["id", "name", "slug", "itemCount"],
        "properties": {
          "id": {
            "type": "string",
            "example": "cat-1"
          },
          "name": {
            "type": "string",
            "example": "Fruits & Vegetables"
          },
          "slug": {
            "type": "string",
            "example": "fruits-vegetables"
          },
          "icon": {
            "type": "string",
            "example": "fruit"
          },
          "itemCount": {
            "type": "integer",
            "example": 154
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "example": "Not Found"
          },
          "message": {
            "type": "string",
            "example": "Product with ID 999 not found"
          },
          "statusCode": {
            "type": "integer",
            "example": 404
          }
        }
      }
    }
  }
}